Fixes post testing
